requirements.txt
# Core Flask
Flask>=3.0.0
Werkzeug>=3.0.0

# Elasticsearch
elasticsearch>=8.11.0,<9.0.0

# MongoDB
pymongo>=4.6.0

# Redis
redis>=5.0.0

# Data Processing (déjà installés avec pip)
pandas>=2.1.0
numpy>=1.26.0

# Fake Data Generation
Faker>=20.0.0

# Environment Variables
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.31.0

# Date/Time
python-dateutil>=2.8.0

# CORS
flask-cors>=4.0.0

# Authentication (pour module intermédiaire)
Flask-Login>=0.6.0

# JWT (pour API)
PyJWT>=2.8.0

# Forms
Flask-WTF>=1.2.0

# API Documentation (Swagger)
flasgger>=0.9.7

# Logging
python-json-logger>=2.0.0

# Testing
pytest>=7.4.0
pytest-flask>=1.3.0
